server/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
gevent==21.8.0
gevent-websocket==0.10.1
Pillow==9.3.0